+++ src/main.ts
@@ -2,4 +2,5 @@
 export { default as Weapon } from './Weapon';
 export { default as Bullet } from './Bullet';
 export { default as consts } from './consts';
+export { default as events } from './events';
 export { default as WeaponPlugin } from './WeaponPlugin';