+++ CHANGELOG.md
@@ -17,6 +17,7 @@ files directly (through a bundler), but prefer to not set up TypeScript compilat
 - Ported plugin to TypeScript
 - Changed exports, the plugin object is now accessible as a named export, or as `WeaponPlugin.WeaponPlugin` as a browser global.
 - The built files now include the necessary polyfills (based on target browsers).
+- Constants previously available at `consts` have been split ut into `consts.KillType`, `consts.Angle` and `consts.FrameType`. They're also available as named exports from `consts.js`.
 
 ### Fixed
 - Fixed several bugs and edge cases, helped by TypeScript