+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2020-04-15
 ### Added
 - Config checks for `trackRotation` and `bulletInheritSpriteSpeed` when following a Sprite without a body or a custom object.
 - Added other build targets, including a legacy build transpiled to ES5 and a modern build supporting the last 2 versions of the major browsers.
@@ -55,6 +57,7 @@ files directly (through a bundler), but prefer to not set up TypeScript compilat
 - Fixed several bugs, like incorrect usage of Rectangle.CenterOn.
 
 [Unreleased]: https://github.com/16patsle/phaser3-weapon-plugin/compare/v1.0.1...HEAD
+[2.0.0]: https://github.com/16patsle/phaser3-weapon-plugin/compare/v1.0.2...v2.0.0
 [1.0.2]: https://github.com/16patsle/phaser3-weapon-plugin/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/16patsle/phaser3-weapon-plugin/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/16patsle/phaser3-weapon-plugin/compare/v1.0.0-beta.1...v1.0.0